LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-lanes/tentative/parsing/grid-lanes-shorthand-valid-expected.txt


FAIL grid-lanes followed by !important assert_equals: expected (string) "\"test\" max-content row" but got (undefined) undefined
FAIL grid-lanes: column fill-reverse "a" calc(10px) should be valid. assert_not_equals: property should be set got disallowed value ""
FAIL grid-lanes: minmax(calc(30% + 5px), 200px) should be valid. assert_not_equals: property should be set got disallowed value ""
FAIL grid-lanes: minmax(10px, 20px) row should be valid. assert_not_equals: property should be set got disallowed value ""
FAIL grid-lanes: 1px 2px should be valid. assert_not_equals: property should be set got disallowed value ""
FAIL grid-lanes: "a" 10px should be valid. assert_not_equals: property should be set got disallowed value ""
FAIL grid-lanes: "a b" 10px 20px row should be valid. assert_not_equals: property should be set got disallowed value ""
FAIL grid-lanes: "a b c" 10% 20% 30% row track-reverse should be valid. assert_not_equals: property should be set got disallowed value ""
FAIL grid-lanes: repeat(5, auto) row should be valid. assert_not_equals: property should be set got disallowed value ""
FAIL grid-lanes: row 10px 20px "a b" should be valid. assert_not_equals: property should be set got disallowed value ""
FAIL grid-lanes: 10px 20px "a b" should be valid. assert_not_equals: property should be set got disallowed value ""
FAIL grid-lanes: column fill-reverse repeat(4, 10px) should be valid. assert_not_equals: property should be set got disallowed value ""
FAIL e.style['grid-lanes'] = "none" should set grid-lanes-direction assert_equals: grid-lanes-direction should be canonical expected (string) "column" but got (undefined) undefined
FAIL e.style['grid-lanes'] = "none" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "none" but got ""
FAIL e.style['grid-lanes'] = "none" should set grid-template-columns assert_equals: grid-template-columns should be canonical expected "none" but got ""
FAIL e.style['grid-lanes'] = "none" should set grid-template-rows assert_equals: grid-template-rows should be canonical expected "none" but got ""
FAIL e.style['grid-lanes'] = "none" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['grid-lanes'] = "10px" should set grid-lanes-direction assert_equals: grid-lanes-direction should be canonical expected (string) "column" but got (undefined) undefined
FAIL e.style['grid-lanes'] = "10px" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "none" but got ""
FAIL e.style['grid-lanes'] = "10px" should set grid-template-columns assert_equals: grid-template-columns should be canonical expected "10px" but got ""
FAIL e.style['grid-lanes'] = "10px" should set grid-template-rows assert_equals: grid-template-rows should be canonical expected "none" but got ""
FAIL e.style['grid-lanes'] = "10px" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['grid-lanes'] = "\"b a\" 20% 40% column" should set grid-lanes-direction assert_equals: grid-lanes-direction should be canonical expected (string) "column" but got (undefined) undefined
FAIL e.style['grid-lanes'] = "\"b a\" 20% 40% column" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "\"b a\"" but got ""
FAIL e.style['grid-lanes'] = "\"b a\" 20% 40% column" should set grid-template-columns assert_equals: grid-template-columns should be canonical expected "20% 40%" but got ""
FAIL e.style['grid-lanes'] = "\"b a\" 20% 40% column" should set grid-template-rows assert_equals: grid-template-rows should be canonical expected "none" but got ""
FAIL e.style['grid-lanes'] = "\"b a\" 20% 40% column" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['grid-lanes'] = "\"b b a\" 1fr 2fr 3fr row" should set grid-lanes-direction assert_equals: grid-lanes-direction should be canonical expected (string) "row" but got (undefined) undefined
FAIL e.style['grid-lanes'] = "\"b b a\" 1fr 2fr 3fr row" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "\"b\" \"b\" \"a\"" but got ""
FAIL e.style['grid-lanes'] = "\"b b a\" 1fr 2fr 3fr row" should set grid-template-columns assert_equals: grid-template-columns should be canonical expected "none" but got ""
FAIL e.style['grid-lanes'] = "\"b b a\" 1fr 2fr 3fr row" should set grid-template-rows assert_equals: grid-template-rows should be canonical expected "1fr 2fr 3fr" but got ""
FAIL e.style['grid-lanes'] = "\"b b a\" 1fr 2fr 3fr row" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['grid-lanes'] = "repeat(2, auto) row track-reverse fill-reverse" should set grid-lanes-direction assert_equals: grid-lanes-direction should be canonical expected (string) "row track-reverse fill-reverse" but got (undefined) undefined
FAIL e.style['grid-lanes'] = "repeat(2, auto) row track-reverse fill-reverse" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "none" but got ""
FAIL e.style['grid-lanes'] = "repeat(2, auto) row track-reverse fill-reverse" should set grid-template-columns assert_equals: grid-template-columns should be canonical expected "none" but got ""
FAIL e.style['grid-lanes'] = "repeat(2, auto) row track-reverse fill-reverse" should set grid-template-rows assert_equals: grid-template-rows should be canonical expected "repeat(2, auto)" but got ""
FAIL e.style['grid-lanes'] = "repeat(2, auto) row track-reverse fill-reverse" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['grid-lanes'] = "\"b a\" 20% 40% normal" should set grid-lanes-direction assert_equals: grid-lanes-direction should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['grid-lanes'] = "\"b a\" 20% 40% normal" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "\"b a\"" but got ""
FAIL e.style['grid-lanes'] = "\"b a\" 20% 40% normal" should set grid-template-columns assert_equals: grid-template-columns should be canonical expected "20% 40%" but got ""
FAIL e.style['grid-lanes'] = "\"b a\" 20% 40% normal" should set grid-template-rows assert_equals: grid-template-rows should be canonical expected "none" but got ""
FAIL e.style['grid-lanes'] = "\"b a\" 20% 40% normal" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['grid-lanes'] = "row fill-reverse 20% 40% \"b a\"" should set grid-lanes-direction assert_equals: grid-lanes-direction should be canonical expected (string) "row fill-reverse" but got (undefined) undefined
FAIL e.style['grid-lanes'] = "row fill-reverse 20% 40% \"b a\"" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "\"b\" \"a\"" but got ""
FAIL e.style['grid-lanes'] = "row fill-reverse 20% 40% \"b a\"" should set grid-template-columns assert_equals: grid-template-columns should be canonical expected "none" but got ""
FAIL e.style['grid-lanes'] = "row fill-reverse 20% 40% \"b a\"" should set grid-template-rows assert_equals: grid-template-rows should be canonical expected "20% 40%" but got ""
FAIL e.style['grid-lanes'] = "row fill-reverse 20% 40% \"b a\"" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['grid-lanes'] = "column track-reverse repeat(5, auto)" should set grid-lanes-direction assert_equals: grid-lanes-direction should be canonical expected (string) "column track-reverse" but got (undefined) undefined
FAIL e.style['grid-lanes'] = "column track-reverse repeat(5, auto)" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "none" but got ""
FAIL e.style['grid-lanes'] = "column track-reverse repeat(5, auto)" should set grid-template-columns assert_equals: grid-template-columns should be canonical expected "repeat(5, auto)" but got ""
FAIL e.style['grid-lanes'] = "column track-reverse repeat(5, auto)" should set grid-template-rows assert_equals: grid-template-rows should be canonical expected "none" but got ""
FAIL e.style['grid-lanes'] = "column track-reverse repeat(5, auto)" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['grid-lanes'] = "\"a b c\" row track-reverse 1fr 2fr 3fr" should set grid-lanes-direction assert_equals: grid-lanes-direction should be canonical expected (string) "row track-reverse" but got (undefined) undefined
FAIL e.style['grid-lanes'] = "\"a b c\" row track-reverse 1fr 2fr 3fr" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "\"a\" \"b\" \"c\"" but got ""
FAIL e.style['grid-lanes'] = "\"a b c\" row track-reverse 1fr 2fr 3fr" should set grid-template-columns assert_equals: grid-template-columns should be canonical expected "none" but got ""
FAIL e.style['grid-lanes'] = "\"a b c\" row track-reverse 1fr 2fr 3fr" should set grid-template-rows assert_equals: grid-template-rows should be canonical expected "1fr 2fr 3fr" but got ""
FAIL e.style['grid-lanes'] = "\"a b c\" row track-reverse 1fr 2fr 3fr" should not set unrelated longhands assert_true: expected true got false
